Refactor the build scripts to support parallel builds, update configuration files, and remove unnecessary code.

+++ misc/clean_git.csc
@@ -1,7 +1,7 @@
 var base_path = system.is_platform_windows() ? ".\\build-cache" : "./build-cache"
 var files = system.path.scan(base_path)
 foreach entry in files
-    if entry.type == system.path.type.dir && entry.name != "." && entry.name != ".."
+    if entry.type == system.path.type.dir && entry.name != "." && entry.name != ".." && system.path.exist(base_path + system.path.separator + entry.name + system.path.separator + ".git")
         system.run("git -C \"" + base_path + system.path.separator + entry.name + "\" clean -dfx")
     end
 end
